refactor into new services folder, separate api services and world generator function
